🚚 Rename: 로딩 컴포넌트 파일 구조 변경 (#227)
